Client-side: move AddressSpaceView cache stats to new scheme
